rename mavis_constants to constants
