+++ mavis/test/data/__init__.py
@@ -7,6 +7,7 @@ import nhs_number
 import pandas as pd
 from faker import Faker
 
+from mavis.test.constants import Programme
 from mavis.test.data_models import (
     Child,
     Clinic,
@@ -14,7 +15,6 @@ from mavis.test.data_models import (
     School,
     User,
 )
-from mavis.test.mavis_constants import Programme
 from mavis.test.utils import (
     get_current_datetime_compact,
     get_current_time_hms_format,
