+++ mavis/test/pages/imports/import_records_wizard_page.py
@@ -5,8 +5,8 @@ from pathlib import Path
 from playwright.sync_api import Page, expect
 
 from mavis.test.annotations import step
+from mavis.test.constants import Programme
 from mavis.test.data import FileMapping, TestData, read_scenario_list_from_file
-from mavis.test.mavis_constants import Programme
 from mavis.test.pages.header_component import HeaderComponent
 from mavis.test.utils import (
     format_datetime_for_upload_link,
