+++ mavis/test/pages/online_consent_wizard_page.py
@@ -3,12 +3,12 @@ from datetime import date
 from playwright.sync_api import Page, expect
 
 from mavis.test.annotations import step
-from mavis.test.data_models import Child, Parent, School
-from mavis.test.mavis_constants import (
+from mavis.test.constants import (
     ConsentOption,
     ConsentRefusalReason,
     Programme,
 )
+from mavis.test.data_models import Child, Parent, School
 
 
 class OnlineConsentWizardPage:
