+++ mavis/test/pages/programmes.py
@@ -5,7 +5,7 @@ import pandas as pd
 from playwright.sync_api import Page
 
 from mavis.test.annotations import step
-from mavis.test.mavis_constants import Child, DeliverySite, Programme, ReportFormat
+from mavis.test.constants import Child, DeliverySite, Programme, ReportFormat
 from mavis.test.pages.header_component import HeaderComponent
 from mavis.test.utils import get_current_datetime_compact
