+++ mavis/test/pages/programmes/programmes_list_page.py
@@ -1,7 +1,7 @@
 from playwright.sync_api import Page
 
 from mavis.test.annotations import step
-from mavis.test.mavis_constants import Programme
+from mavis.test.constants import Programme
 from mavis.test.pages.header_component import HeaderComponent
