+++ mavis/test/pages/reports/reports_vaccinations_page.py
@@ -6,7 +6,7 @@ import requests
 from playwright.sync_api import Page, expect
 
 from mavis.test.annotations import step
-from mavis.test.mavis_constants import Programme
+from mavis.test.constants import Programme
 from mavis.test.pages.header_component import HeaderComponent
 from mavis.test.pages.reports.reports_tabs import ReportsTabs
