+++ mavis/test/pages/school_moves/download_school_moves_page.py
@@ -6,8 +6,8 @@ from pandas import DataFrame, Series
 from playwright.sync_api import Page
 
 from mavis.test.annotations import step
+from mavis.test.constants import SCHOOL_MOVE_HEADERS
 from mavis.test.data_models import Child, School
-from mavis.test.mavis_constants import SCHOOL_MOVE_HEADERS
 from mavis.test.pages.header_component import HeaderComponent
