+++ mavis/test/pages/sessions/sessions_patient_page.py
@@ -3,16 +3,16 @@ import time
 from playwright.sync_api import Page, expect
 
 from mavis.test.annotations import step
+from mavis.test.constants import (
+    ConsentOption,
+    DeliverySite,
+    Programme,
+)
 from mavis.test.data_models import (
     Child,
     Parent,
     School,
     VaccinationRecord,
-)
-from mavis.test.mavis_constants import (
-    ConsentOption,
-    DeliverySite,
-    Programme,
 )
 from mavis.test.pages.header_component import HeaderComponent
 from mavis.test.utils import (
