+++ mavis/test/pages/vaccination_record/edit_vaccination_record_page.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from playwright.sync_api import Page, expect
 
 from mavis.test.annotations import step
-from mavis.test.mavis_constants import DeliverySite
+from mavis.test.constants import DeliverySite
 from mavis.test.pages.header_component import HeaderComponent
