+++ tests/test_children.py
@@ -1,9 +1,9 @@
 import pytest
 
 from mavis.test.annotations import issue
+from mavis.test.constants import Programme
 from mavis.test.data import ClassFileMapping, CohortsFileMapping, VaccsFileMapping
 from mavis.test.helpers.accessibility_helper import AccessibilityHelper
-from mavis.test.mavis_constants import Programme
 from mavis.test.pages import (
     ChildActivityLogPage,
     ChildArchivePage,
