+++ tests/test_consent_responses.py
@@ -2,9 +2,9 @@ import pytest
 from playwright.sync_api import expect
 
 from mavis.test.annotations import issue
+from mavis.test.constants import Programme
 from mavis.test.data import ClassFileMapping, CohortsFileMapping, pds
 from mavis.test.helpers.accessibility_helper import AccessibilityHelper
-from mavis.test.mavis_constants import Programme
 from mavis.test.pages import (
     ArchiveConsentResponsePage,
     ChildActivityLogPage,
