+++ tests/test_e2e_doubles.py
@@ -1,7 +1,7 @@
 import pytest
 
+from mavis.test.constants import Programme, Vaccine
 from mavis.test.data_models import VaccinationRecord
-from mavis.test.mavis_constants import Programme, Vaccine
 from mavis.test.pages import (
     DashboardPage,
     LogInPage,
