+++ tests/test_e2e_flu.py
@@ -1,12 +1,12 @@
 import pytest
 
 from mavis.test.annotations import issue
-from mavis.test.data_models import VaccinationRecord
-from mavis.test.mavis_constants import (
+from mavis.test.constants import (
     ConsentOption,
     Programme,
     Vaccine,
 )
+from mavis.test.data_models import VaccinationRecord
 from mavis.test.pages import (
     ChildRecordPage,
     ChildrenSearchPage,
