+++ tests/test_import_class_lists.py
@@ -1,7 +1,7 @@
 import pytest
 
+from mavis.test.constants import Programme
 from mavis.test.data import ClassFileMapping
-from mavis.test.mavis_constants import Programme
 from mavis.test.pages import (
     ChildrenSearchPage,
     DashboardPage,
