+++ tests/test_import_offline_vaccinations.py
@@ -1,8 +1,8 @@
 import pytest
 
 from mavis.test.annotations import issue
+from mavis.test.constants import Programme
 from mavis.test.data import ClassFileMapping, VaccsFileMapping
-from mavis.test.mavis_constants import Programme
 from mavis.test.pages import (
     ChildRecordPage,
     ChildrenSearchPage,
