+++ tests/test_nurse_consent.py
@@ -1,9 +1,9 @@
 import pytest
 
 from mavis.test.annotations import issue
+from mavis.test.constants import ConsentMethod, DeliverySite, Programme, Vaccine
 from mavis.test.data import CohortsFileMapping
 from mavis.test.helpers.accessibility_helper import AccessibilityHelper
-from mavis.test.mavis_constants import ConsentMethod, DeliverySite, Programme, Vaccine
 from mavis.test.pages import (
     DashboardPage,
     GillickCompetencePage,
