+++ tests/test_online_consent_flu.py
@@ -1,8 +1,8 @@
 import pytest
 
 from mavis.test.annotations import issue
+from mavis.test.constants import ConsentOption, ConsentRefusalReason, Programme
 from mavis.test.data import CohortsFileMapping
-from mavis.test.mavis_constants import ConsentOption, ConsentRefusalReason, Programme
 from mavis.test.pages import (
     DashboardPage,
     ImportRecordsWizardPage,
