+++ tests/test_online_consent_hpv.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mavis.test.mavis_constants import ConsentRefusalReason, Programme
+from mavis.test.constants import ConsentRefusalReason, Programme
 from mavis.test.pages import (
     OnlineConsentWizardPage,
     StartPage,
