+++ tests/test_online_consent_school_moves.py
@@ -1,8 +1,8 @@
 import pytest
 
+from mavis.test.constants import ConsentOption, Programme
 from mavis.test.data import CohortsFileMapping
 from mavis.test.helpers.accessibility_helper import AccessibilityHelper
-from mavis.test.mavis_constants import ConsentOption, Programme
 from mavis.test.pages import (
     ConsentResponsePage,
     CreateNewRecordConsentResponsePage,
