+++ tests/test_programmes.py
@@ -1,9 +1,9 @@
 import pytest
 
 from mavis.test.annotations import issue
+from mavis.test.constants import Programme, ReportFormat
 from mavis.test.data import CohortsFileMapping
 from mavis.test.helpers.accessibility_helper import AccessibilityHelper
-from mavis.test.mavis_constants import Programme, ReportFormat
 from mavis.test.pages import (
     ChildArchivePage,
     ChildRecordPage,
