+++ tests/test_reporting.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mavis.test.mavis_constants import (
+from mavis.test.constants import (
     Programme,
 )
 from mavis.test.pages import (
