+++ tests/test_school_moves.py
@@ -1,9 +1,9 @@
 import pytest
 from playwright.sync_api import expect
 
+from mavis.test.constants import Programme
 from mavis.test.data import ClassFileMapping
 from mavis.test.helpers.accessibility_helper import AccessibilityHelper
-from mavis.test.mavis_constants import Programme
 from mavis.test.pages import (
     ChildActivityLogPage,
     ChildRecordPage,
