+++ tests/test_tallying.py
@@ -1,16 +1,16 @@
 import pytest
 
 from mavis.test.annotations import issue
-from mavis.test.data import ClassFileMapping
-from mavis.test.mavis_constants import (
+from mavis.test.constants import (
     ConsentMethod,
     ConsentOption,
     ConsentRefusalReason,
     Programme,
     TallyCategory,
-    VaccinationRecord,
     Vaccine,
 )
+from mavis.test.data import ClassFileMapping
+from mavis.test.data_models import VaccinationRecord
 from mavis.test.pages import (
     DashboardPage,
     ImportRecordsWizardPage,
